Happy Holiday! w/ New Input mode to salmon

+++ include/SalmonUtils.hpp
@@ -115,7 +115,7 @@ TranscriptGeneMap
 transcriptToGeneMapFromFasta(const std::string& transcriptsFile);
 
 bool readEquivCounts(boost::filesystem::path& eqFilePathString, std::vector<string>& tnames,
-                     std::vector<std::vector<uint64_t>>& eqclasses,
+                     std::vector<std::vector<uint32_t>>& eqclasses,
                      std::vector<std::vector<double>>& auxs_vals,
                      std::vector<uint32_t>& eqclass_counts );
